added role based access to feedbackEndpoints

+++ HMS_API/Endpoints/FeedbackEndpoints.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Security.Claims;
 using HMS_API.Data;
 using HMS_API.Dtos.feedback;
 using HMS_API.Entities;
@@ -25,13 +26,16 @@ public class FeedbackEndpoints
     public RouteGroupBuilder MapFeedbackEndpoints(WebApplication app)
     {
         var feedback = app.MapGroup("feedback")
-                          .WithParameterValidation();
+                          .WithParameterValidation()
+                          .RequireAuthorization();
 
         // Requests for feedback
-        feedback.MapGet("/", async (HMS_Context db) => 
+        feedback.MapGet("/", async (HMS_Context db, ClaimsPrincipal account) => 
         {
             try
             {
+                if(VerifyAdminStudentClaim(account)== false) // Check if the user has the Admin or Lecturer role using HasClaim
+                    return Results.Forbid(); // Return 403 Forbidden if neither role
                 var feedbacks = await db.Feedbacks
                     .Include(f => f.Submission) // Include submission entity to get info as in DTO
                     .Include(f => f.UserAccount) // Include user entity to get info as in DTO
@@ -47,13 +51,16 @@ public class FeedbackEndpoints
                 logger.LogError(ex, "Failed to fetch feedbacks/ Date/Time: {dateTime}.", DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss"));
                 return Results.Problem("An error occurred while retrieving feedbacks.");
             }
-        });
+        }).RequireAuthorization();
 
         // Get specific feedback
-        feedback.MapGet("/{id}", async (int id, HMS_Context db) => 
+        feedback.MapGet("/{id}", async (int id, HMS_Context db, ClaimsPrincipal account) => 
         {
             try
             {
+                if(VerifyAdminStudentClaim(account) == false) // Check if the user has the Admin or student role using HasClaim
+                    return Results.Forbid(); // Return 403 Forbidden if neither role
+
                 Feedback? feedback = await db.Feedbacks.FindAsync(id);
 
                 return feedback is null ? Results.NotFound() : Results.Ok(feedback.ToFeedbackDetailsDto());
@@ -63,13 +70,16 @@ public class FeedbackEndpoints
                 logger.LogError(ex, "Failed to fetch feedback with ID {Id}/ Date/Time: {dateTime}.", id, DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss"));
                 return Results.Problem("An error occurred while retrieving feedback.");
             }
-        }).WithName(GetFeedbackEndpointName);
+        }).RequireAuthorization().WithName(GetFeedbackEndpointName);
 
         // Post feedback
-        feedback.MapPost("/", async (CreateFeedbackDto newFeedback, HMS_Context db) => 
+        feedback.MapPost("/", async (CreateFeedbackDto newFeedback, HMS_Context db, ClaimsPrincipal account) => 
         {
             try
             {
+                if(VerifyAdminLecturerClaim(account)== false) // Check if the user has the Admin or Lecturer role using HasClaim
+                    return Results.Forbid(); // Return 403 Forbidden if neither role
+
                 Feedback feedback = newFeedback.ToEntity();
                 // force date creation
                 DateOnly currentDate = DateOnly.FromDateTime(DateTime.UtcNow);
@@ -88,18 +98,21 @@ public class FeedbackEndpoints
                 logger.LogError(ex, "Failed to create feedback.");
                 return Results.Problem("An error occurred while creating feedback.");
             }
-        }).WithParameterValidation();
+        }).RequireAuthorization().WithParameterValidation();
 
         // Put = update feedback
-        feedback.MapPut("/{id}", async (int id, UpdateFeedbackDto updateFeedback, HMS_Context db) => 
+        feedback.MapPut("/{id}", async (int id, UpdateFeedbackDto updateFeedback, HMS_Context db, ClaimsPrincipal account) => 
         {
-                var existingFeedback = await db.Feedbacks.FindAsync(id);
+            if(VerifyAdminLecturerClaim(account)== false) // Check if the user has the Admin or Lecturer role using HasClaim
+                return Results.Forbid(); // Return 403 Forbidden if neither role
 
-                if (existingFeedback == null) // Feedback must exst
-                {
-                    logger.LogWarning("Feedback with ID {id} not found for update/ Date/Time: {dateTime}.", id, DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss"));
-                    return Results.NotFound();
-                }
+            var existingFeedback = await db.Feedbacks.FindAsync(id);
+
+            if (existingFeedback == null) // Feedback must exist
+            {
+                logger.LogWarning("Feedback with ID {id} not found for update/ Date/Time: {dateTime}.", id, DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss"));
+                return Results.NotFound();
+            }
             try
             {
                 // inject updated information
@@ -121,13 +134,16 @@ public class FeedbackEndpoints
                 logger.LogError(ex, "Failed to update feedback with ID {Id}/ Date/Time: {dateTime}.", id, DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss"));
                 return Results.Problem("An error occurred while updating feedback.");
             }
-        });
+        }).RequireAuthorization();
 
         // Delete feedback
-        feedback.MapDelete("/{id}", async (int id, HMS_Context db) => 
+        feedback.MapDelete("/{id}", async (int id, HMS_Context db, ClaimsPrincipal account) => 
         {
             try
             {
+                if(VerifyAdminLecturerClaim(account)== false) // Check if the user has the Admin or Lecturer role using HasClaim
+                return Results.Forbid(); // Return 403 Forbidden if neither role
+
                 var rowsAffected = await db.Feedbacks.Where(f => f.Id == id).ExecuteDeleteAsync();
                 if (rowsAffected == 0)
                 {
@@ -143,8 +159,30 @@ public class FeedbackEndpoints
                 logger.LogError(ex, "An error occurred while deleting feedback with ID {Id}/ Date/Time: {dateTime}.", id, DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss"));
                 return Results.Problem("An error occurred while deleting feedback.");
             }
-        });
+        }).RequireAuthorization();
 
         return feedback;
+    }
+    private bool VerifyAdminLecturerClaim(ClaimsPrincipal user) // verify admin or lecturer
+    {
+        bool verify = false;
+        if (user.HasClaim(claim => 
+            claim.Type == "http://schemas.microsoft.com/ws/2008/06/identity/claims/role" && 
+                (claim.Value == "Admin" || claim.Value == "Lecturer")))
+                {
+                    verify = true; 
+                }
+        return verify;
+    }
+    private bool VerifyAdminStudentClaim(ClaimsPrincipal user) // verify admin or student
+    {
+        bool verify = false;
+        if (user.HasClaim(claim => 
+            claim.Type == "http://schemas.microsoft.com/ws/2008/06/identity/claims/role" && 
+                (claim.Value == "Admin" || claim.Value == "Student")))
+                {
+                    verify = true; 
+                }
+        return verify;
     }
 }
